+++ .gitignore
@@ -19,6 +19,7 @@ lib64/
 parts/
 sdist/
 var/
+./src/*
 wheels/
 share/python-wheels/
 *.egg-info/
