+++ MANIFEST.in
@@ -1,0 +1,4 @@
+include LICENSE
+include README.md
+
+recursive-include trips/django_trips/ *
