--- django_trips/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE=django_trips.settings
-python_files = test_*.py
