--- django_trips/trips/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TripsConfig(AppConfig):
-    name = 'trips'
