--- django_trips/trips/mixins.py
@@ -1,83 +1,0 @@
-"""Model mixins"""
-from django.utils.text import slugify
-import re
-
-
-class SlugMixin:
-    def create_slug(self, slug_field="name", *args, **kwargs):
-        """ long stuff to ensure the slug is unique """
-        unique_slugify(self, getattr(self, slug_field))
-
-    def save(self, *args, **kwargs):
-        if any([self.pk is None, self.slug is None]):
-            self.create_slug(*args, **kwargs)
-        return super().save(*args, **kwargs)
-
-
-def unique_slugify(instance, value, slug_field_name='slug', queryset=None,
-                   slug_separator='-'):
-    """
-    Calculates and stores a unique slug of ``value`` for an instance.
-
-    ``slug_field_name`` should be a string matching the name of the field to
-    store the slug in (and the field to check against for uniqueness).
-
-    ``queryset`` usually doesn't need to be explicitly provided - it'll default
-    to using the ``.all()`` queryset from the model's default manager.
-    """
-    slug_field = instance._meta.get_field(slug_field_name)
-
-    slug_len = slug_field.max_length
-
-    # Sort out the initial slug, limiting its length if necessary.
-    slug = slugify(value)
-    if slug_len:
-        slug = slug[:slug_len]
-    slug = _slug_strip(slug, slug_separator)
-    original_slug = slug
-
-    # Create the queryset if one wasn't explicitly provided and exclude the
-    # current instance from the queryset.
-    if queryset is None:
-        queryset = instance.__class__._default_manager.all()
-    if instance.pk:
-        queryset = queryset.exclude(pk=instance.pk)
-
-    # Find a unique slug. If one matches, at '-2' to the end and try again
-    # (then '-3', etc).
-    next = 2
-    while not slug or queryset.filter(**{slug_field_name: slug}):
-        slug = original_slug
-        end = '%s%s' % (slug_separator, next)
-        if slug_len and len(slug) + len(end) > slug_len:
-            slug = slug[:slug_len - len(end)]
-            slug = _slug_strip(slug, slug_separator)
-        slug = '%s%s' % (slug, end)
-        next += 1
-
-    setattr(instance, slug_field.attname, slug)
-
-
-def _slug_strip(value, separator='-'):
-    """
-    Cleans up a slug by removing slug separator characters that occur at the
-    beginning or end of a slug.
-
-    If an alternate separator is used, it will also replace any instances of
-    the default '-' separator with the new separator.
-    """
-    separator = separator or ''
-    if separator == '-' or not separator:
-        re_sep = '-'
-    else:
-        re_sep = '(?:-|%s)' % re.escape(separator)
-    # Remove multiple instances and if an alternate separator is provided,
-    # replace the default '-' separator.
-    if separator != re_sep:
-        value = re.sub('%s+' % re_sep, separator, value)
-    # Remove separator from the beginning and end of the slug.
-    if separator:
-        if separator != '-':
-            re_sep = re.escape(separator)
-        value = re.sub(r'^%s+|%s+$' % (re_sep, re_sep), '', value)
-    return value
