--- django_trips/trips/urls.py
@@ -1,6 +1,0 @@
-from django.conf.urls import url
-from django.views import generic
-
-urlpatterns = [
-
-]
