+++ setup.cfg
@@ -1,0 +1,28 @@
+[metadata]
+name = django-trips
+version = 0.2.7.1
+description = A Django Rest API for fetching and creating trips and their schedules.
+long_description = file: README.md
+url = https://github.com/awaisdar001/django-trips
+author = Awais Jibran
+author_email = awaisdar001@gmail.com
+license = MIT License
+classifiers =
+  Environment :: Web Environment
+  Framework :: Django
+  Framework :: Django :: 2.2
+  Intended Audience :: Developers
+  License :: OSI Approved :: MIT License
+  Operating System :: OS Independent
+  Programming Language :: Python
+  Programming Language :: Python :: 3
+  Programming Language :: Python :: 3 :: Only
+  Programming Language :: Python :: 3.6
+  Programming Language :: Python :: 3.7
+  Programming Language :: Python :: 3.8
+  Topic :: Internet :: WWW/HTTP
+  Topic :: Internet :: WWW/HTTP :: Dynamic Content
+
+[options]
+include_package_data = true
+packages = find:
