+++ trips/django_trips/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DjangoTripsConfig(AppConfig):
+    name = 'django trips'
