+++ trips/django_trips/managers.py
