+++ trips/django_trips/tests.py
