+++ trips/django_trips/urls.py
@@ -1,0 +1,9 @@
+"""Urls for trips app"""
+from django.conf.urls import include
+from django.urls import path
+
+app_label="ol"
+urlpatterns = [
+    path('api/', include('django_trips.api.urls')),
+
+]
