+++ trips/django_trips/views.py
@@ -1,0 +1,1 @@
+""" - """
