+++ trips/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+DJANGO_SETTINGS_MODULE=trips.settings
+python_files = test_*.py
